els-ASAP auto commit file/tag perftest-2.0-2.el6.src.rpm

+++ src/src/multicast_resources.h
@@ -12,7 +12,7 @@
  *     conditions are met:
  *
  *      - Redistributions of source code must retain the above
- *        copyright notice, this list of conditions and the following						    			    
+ *        copyright notice, this list of conditions and the following
  *        disclaimer.
  *
  *      - Redistributions in binary form must reproduce the above
@@ -32,33 +32,32 @@
  * Author: Ido Shamay <idos@dev.mellanox.co.il>
  */
 
-
 #ifndef MULTICAST_RESOURCES_H
 #define MULTICAST_RESOURCES_H
 
  /* Multicast Module for perftest.
   *
-  * Description : 
+  * Description :
   *
-  *   This file contains the structures and methods for implementing a multiple 
+  *   This file contains the structures and methods for implementing a multiple
   *   multicast groups in user space enviroment.
   *	  The module is in use in "send_bw" and "send_lat" ,but can be used on other
   *	  applications and can generate more methods and serve more benchmarks.
   *   The Module uses only the structire defined here , enabling generic use of it.
-  *   
+  *
   * Defined Types :
   *
   *   mcast_parameters - Contains all the parameters needed for this module.
   *   mcast_group      - The multicast group entitiy itself.
   *   mcg_qp		   - Is a QP structure that is attahced to the group.
-  *   
+  *
   */
 
 
-/************************************************************************ 
+/************************************************************************
  *   Macros , Defines and Files included for work.	    			    *
  ************************************************************************/
- 
+
 #include <infiniband/verbs.h>
 #include <infiniband/umad.h>
 #include "get_clock.h"
@@ -70,7 +69,7 @@
 #define MAX_POLL_ITERATION_TIMEOUT  1000000
 #define MCG_GID {255,1,0,0,0,2,201,133,0,0,0,0,0,0,0,0}
 
-//  Definitions section for MADs 
+//  Definitions section for MADs
 #define SUBN_ADM_ATTR_MC_MEMBER_RECORD 0x38
 #define MANAGMENT_CLASS_SUBN_ADM       0x03 	  /* Subnet Administration class */
 #define MCMEMBER_JOINSTATE_FULL_MEMBER 0x1
@@ -80,20 +79,20 @@
 #define DEF_TCLASS                     0
 #define DEF_FLOW_LABLE                 0
 
-// Macro for 64 bit variables to switch to from net 
-#define ntohll(x) (((u_int64_t)(ntohl((int)((x << 32) >> 32))) << 32) | (unsigned int)ntohl(((int)(x >> 32)))) 
+// Macro for 64 bit variables to switch to from net
+#define ntohll(x) (((uint64_t)(ntohl((int)((x << 32) >> 32))) << 32) | (unsigned int)ntohl(((int)(x >> 32))))
 #define htonll(x) ntohll(x)
 
-// generate a bit mask S bits width 
-#define MASK32(S)  ( ((u_int32_t) ~0L) >> (32-(S)) )
+// generate a bit mask S bits width
+#define MASK32(S)  ( ((uint32_t) ~0L) >> (32-(S)) )
 
 // generate a bit mask with bits O+S..O set (assumes 32 bit integer).
 #define BITS32(O,S) ( MASK32(S) << (O) )
 
-// extract S bits from (u_int32_t)W with offset O and shifts them O places to the right 
+// extract S bits from (u_int32_t)W with offset O and shifts them O places to the right
 #define EXTRACT32(W,O,S) ( ((W)>>(O)) & MASK32(S) )
 
-// insert S bits with offset O from field F into word W (u_int32_t) 
+// insert S bits with offset O from field F into word W (u_int32_t)
 #define INSERT32(W,F,O,S) (/*(W)=*/ ( ((W) & (~BITS32(O,S)) ) | (((F) & MASK32(S))<<(O)) ))
 
 #ifndef INSERTF
@@ -101,7 +100,7 @@
 #endif
 
 
-// according to Table 187 in the IB spec 1.2.1 
+// according to Table 187 in the IB spec 1.2.1
 typedef enum {
 	SUBN_ADM_METHOD_SET    = 0x2,
 	SUBN_ADM_METHOD_DELETE = 0x15
@@ -125,7 +124,7 @@ typedef enum {
 } mcast_state;
 
 
-/************************************************************************ 
+/************************************************************************
  *   Multicast data structures.						    			    *
  ************************************************************************/
 
@@ -135,7 +134,7 @@ struct mcast_parameters {
 	int					  is_user_mgid;
 	int					  mcast_state;
 	int 				  ib_port;
-	int 				  mlid;
+	uint16_t			  mlid;
 	const char			  *user_mgid;
 	const char			  *ib_devname;
 	uint16_t 			  pkey;
@@ -143,10 +142,11 @@ struct mcast_parameters {
 	uint8_t 			  sm_sl;
 	union ibv_gid 		  port_gid;
 	union ibv_gid 		  mgid;
-	
+
 };
 
-// according to Table 195 in the IB spec 1.2.1 
+// according to Table 195 in the IB spec 1.2.1
+
 struct sa_mad_packet_t {
 	u_int8_t		mad_header_buf[24];
 	u_int8_t		rmpp_header_buf[12];
@@ -157,7 +157,7 @@ struct sa_mad_packet_t {
 	u_int8_t		SubnetAdminData[200];
 }__attribute__((packed));
 
-/************************************************************************ 
+/************************************************************************
  *   Multicast resources methods.					    			    *
  ************************************************************************/
 
@@ -165,12 +165,12 @@ struct sa_mad_packet_t {
  *
  * Description :
  *
- *  Sets the Multicast GID , and stores it in the "mgid" value of 
+ *  Sets the Multicast GID , and stores it in the "mgid" value of
  *  mcast resourcs. If the user requested for a specific MGID, which
  *  is stored in params->user_mgid (in this case params->is_user_mgid should be 1)
- *  than it will be his MGID, if not the library choose a default one.  
+ *  than it will be his MGID, if not the library choose a default one.
  *
- * Parameters : 
+ * Parameters :
  *
  *  params            - The parameters of the machine
  *  my_dest ,rem_dest - The 2 sides that ends the connection.
@@ -187,7 +187,7 @@ void set_multicast_gid(struct mcast_parameters *params,uint32_t qp_num,int is_cl
  *  Close the connection between the 2 machines.
  *  It performs an handshake to ensure the 2 sides are there.
  *
- * Parameters : 
+ * Parameters :
  *
  *  params            - The parameters of the machine
  *  my_dest ,rem_dest - The 2 sides that ends the connection.
